nvim: suppress undefined-global warnings for "vim"
vim global gets defined by nvim's Lua interpreter automatically when the
init.lua is loaded. Because of this, it is undefined as far as lua
language server is concerned, and is reported as a warning through out
the file.

To avoid having a bunch o spurious warnings, and thus risk missing
non-spurious ones, this needs to be suppressed. Simply disabling
undefined-global for the entire file will risk missing genuinely
undefined symbols (e.g. due to a typo).

So instead, create a local definition for "vim" that just assigns the
"non-existing" global, and then suppress undefined-global just for that
line.